+++ changelog.md
@@ -81,3 +81,10 @@ update:
 fix:
     1. fix a bug that versionTag didn't display in page
     2. fix a bug that delete a project that mongodb cause server crash
+    
+v1.9.0_test
+add:
+    1. support drag and drop local pro folder to add it to web server 
+update:
+fix:
+    1.fix a bug that cause upload a folder who is a download local project fail